backend/requirements.txt
Django==5.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
python-decouple
